Move projectiles to their own namespace and folder
While they are (currently) technically effects, this makes the actual purpose and wiki more clear.
