+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -147,21 +147,21 @@
     <Compile Include="AI\States\StateBase.cs" />
     <Compile Include="AI\SupportPowerDecision.cs" />
     <Compile Include="Effects\Beacon.cs" />
-    <Compile Include="Effects\Bullet.cs" />
     <Compile Include="Effects\Contrail.cs" />
     <Compile Include="Effects\ContrailFader.cs" />
     <Compile Include="Effects\CrateEffect.cs" />
     <Compile Include="Effects\FlashTarget.cs" />
     <Compile Include="Effects\FloatingText.cs" />
-    <Compile Include="Effects\GravityBomb.cs" />
-    <Compile Include="Effects\LaserZap.cs" />
-    <Compile Include="Effects\AreaBeam.cs" />
-    <Compile Include="Effects\Missile.cs" />
     <Compile Include="Effects\NukeLaunch.cs" />
     <Compile Include="Effects\PowerdownIndicator.cs" />
     <Compile Include="Effects\RallyPointIndicator.cs" />
     <Compile Include="Effects\RepairIndicator.cs" />
     <Compile Include="Effects\SpriteEffect.cs" />
+    <Compile Include="Projectiles\AreaBeam.cs" />
+    <Compile Include="Projectiles\Bullet.cs" />
+    <Compile Include="Projectiles\GravityBomb.cs" />
+    <Compile Include="Projectiles\LaserZap.cs" />
+    <Compile Include="Projectiles\Missile.cs" />
     <Compile Include="Commands\ChatCommands.cs" />
     <Compile Include="Commands\DevCommands.cs" />
     <Compile Include="Commands\HelpCommand.cs" />
