+++ OpenRA.Mods.Common/Projectiles/GravityBomb.cs
@@ -13,9 +13,10 @@ using System.Collections.Generic;
 using OpenRA.Effects;
 using OpenRA.GameRules;
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Effects;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Common.Effects
+namespace OpenRA.Mods.Common.Projectiles
 {
 	public class GravityBombInfo : IProjectileInfo
 	{
